# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for SUT SE 2563",
        "title": "SUT SA Example API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/paytype": {
            "get": {
                "description": "list paytype entities",
                "produces": [
                    "application/json"
                ],
                "summary": "List paytype entities",
                "operationId": "list-paytype",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/ent.PayType"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "post": {
                "description": "Create paytype",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create paytype",
                "operationId": "create-paytype",
                "parameters": [
                    {
                        "description": "PayType entity",
                        "name": "paytype",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.PayType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.PayType"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
        "/paytype/{id}": {
            "get": {
                "description": "get paytype by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Get a paytype entity by ID",
                "operationId": "get-paytype",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PayType ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.PayType"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "put": {
                "description": "update paytype by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update a paytype entity by ID",
                "operationId": "update-paytype",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PayType ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "PayType entity",
                        "name": "paytype",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ent.PayType"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ent.PayType"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            },
            "delete": {
                "description": "get paytype by ID",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete a paytype entity by ID",
                "operationId": "delete-paytype",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "PayType ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/gin.H"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "ent.Bill": {
            "type": "object",
            "properties": {
                "Amount": {
                    "description": "Amount holds the value of the \"Amount\" field.",
                    "type": "integer"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the BillQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BillEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.BillEdges": {
            "type": "object",
            "properties": {
                "billstatus": {
                    "description": "Billstatus holds the value of the billstatus edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BillStatus"
                },
                "payments": {
                    "description": "Payments holds the value of the payments edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Payment"
                },
                "resident": {
                    "description": "Resident holds the value of the resident edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.User"
                }
            }
        },
        "ent.BillStatus": {
            "type": "object",
            "properties": {
                "BillStatus": {
                    "description": "BillStatus holds the value of the \"BillStatus\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the BillStatusQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.BillStatusEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.BillStatusEdges": {
            "type": "object",
            "properties": {
                "bills": {
                    "description": "Bills holds the value of the bills edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Bill"
                    }
                }
            }
        },
        "ent.PayType": {
            "type": "object",
            "properties": {
                "TypeInfo": {
                    "description": "TypeInfo holds the value of the \"TypeInfo\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PayTypeQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PayTypeEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.PayTypeEdges": {
            "type": "object",
            "properties": {
                "payments": {
                    "description": "Payments holds the value of the payments edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Payment"
                    }
                }
            }
        },
        "ent.Payment": {
            "type": "object",
            "properties": {
                "AddDatetime": {
                    "description": "AddDatetime holds the value of the \"AddDatetime\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PaymentQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PaymentEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.PaymentEdges": {
            "type": "object",
            "properties": {
                "bill": {
                    "description": "Bill holds the value of the bill edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.Bill"
                },
                "paymentstatus": {
                    "description": "Paymentstatus holds the value of the paymentstatus edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PaymentStatus"
                },
                "paytype": {
                    "description": "Paytype holds the value of the paytype edge.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PayType"
                }
            }
        },
        "ent.PaymentStatus": {
            "type": "object",
            "properties": {
                "PaymentStatus": {
                    "description": "PaymentStatus holds the value of the \"PaymentStatus\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the PaymentStatusQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.PaymentStatusEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.PaymentStatusEdges": {
            "type": "object",
            "properties": {
                "payments": {
                    "description": "Payments holds the value of the payments edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Payment"
                    }
                }
            }
        },
        "ent.User": {
            "type": "object",
            "properties": {
                "Email": {
                    "description": "Email holds the value of the \"Email\" field.",
                    "type": "string"
                },
                "Name": {
                    "description": "Name holds the value of the \"Name\" field.",
                    "type": "string"
                },
                "edges": {
                    "description": "Edges holds the relations/edges for other nodes in the graph.\nThe values are being populated by the UserQuery when eager-loading is set.",
                    "type": "object",
                    "$ref": "#/definitions/ent.UserEdges"
                },
                "id": {
                    "description": "ID of the ent.",
                    "type": "integer"
                }
            }
        },
        "ent.UserEdges": {
            "type": "object",
            "properties": {
                "bills": {
                    "description": "Bills holds the value of the bills edge.",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ent.Bill"
                    }
                }
            }
        },
        "gin.H": {
            "type": "object",
            "additionalProperties": true
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "BasicAuth": {
            "type": "basic"
        },
        "OAuth2AccessCode": {
            "type": "oauth2",
            "flow": "accessCode",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information"
            }
        },
        "OAuth2Application": {
            "type": "oauth2",
            "flow": "application",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Implicit": {
            "type": "oauth2",
            "flow": "implicit",
            "authorizationUrl": "https://example.com/oauth/authorize",
            "scopes": {
                "admin": " Grants read and write access to administrative information",
                "write": " Grants write access"
            }
        },
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "https://example.com/oauth/token",
            "scopes": {
...
}